Move shared UIActions which open or activate a tool form to new static class ToolForms.cs.
